backend/backend_requirements.txt
# PursuitDocs Backend Dependencies

# Web framework
fastapi
mangum
python-multipart
uvicorn

# LangGraph + LangChain
langgraph
langchain-anthropic

# Vector store
chromadb
openai
tiktoken

# PDF extraction
pdfplumber

# HTML parsing
beautifulsoup4
requests

# Utilities
python-dotenv
langdetect

# Document export
python-docx

# AWS (S3 for Chroma + firm profiles, DynamoDB for rate limiting)
boto3
